+++ src/module_writer.rs
@@ -17,9 +17,9 @@ use std::io::{Read, Write};
 #[cfg(not(target_os = "windows"))]
 use std::os::unix::fs::OpenOptionsExt;
 use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::process::{Command, Output};
 use std::str;
-use tempfile::tempdir;
+use tempfile::{tempdir, TempDir};
 use walkdir::WalkDir;
 use zip::{self, ZipWriter};
 
@@ -339,23 +339,26 @@ del os
 "#
 }
 
-/// Returns the content of what will become ffi.py by invocing cbindgen and cffi
-///
-/// First we check if user has provided their own header at `target/header.h`, otherwise
-/// we run cbindgen to generate onw.
-///
-/// We're using the cffi recompiler, which reads the header, translates them into instructions
-/// how to load the shared library without the header and then writes those instructions to a
-/// file called `ffi.py`. This `ffi.py` will expose an object called `ffi`. This object is used
-/// in `__init__.py` to load the shared library into a module called `lib`.
-pub fn generate_cffi_declarations(crate_dir: &Path, python: &PathBuf) -> Result<String> {
-    let tempdir = tempdir()?;
+/// Wraps some boilerplate around error handling when calling python
+fn call_python<I, S>(python: &Path, args: I) -> Result<Output>
+where
+    I: IntoIterator<Item = S>,
+    S: AsRef<OsStr>,
+{
+    Ok(Command::new(&python)
+        .args(args)
+        .output()
+        .context(format!("Failed to run python at {:?}", &python))?)
+}
+
+/// Checks if user has provided their own header at `target/header.h`, otherwise
+/// we run cbindgen to generate one.
+fn cffi_header(crate_dir: &Path, tempdir: &TempDir) -> Result<PathBuf> {
     let maybe_header = crate_dir.join("target").join("header.h");
 
-    let header;
     if maybe_header.is_file() {
         println!("💼 Using the existing header at {}", maybe_header.display());
-        header = maybe_header;
+        Ok(maybe_header)
     } else {
         if crate_dir.join("cbindgen.toml").is_file() {
             println!(
@@ -380,15 +383,30 @@ pub fn generate_cffi_declarations(crate_dir: &Path, python: &PathBuf) -> Result<
             .generate()
             .context("Failed to run cbindgen")?;
 
-        header = tempdir.as_ref().join("header.h");
+        let header = tempdir.as_ref().join("header.h");
         bindings.write_to_file(&header);
+        Ok(header)
     }
+}
+
+/// Returns the content of what will become ffi.py by invoking cbindgen and cffi
+///
+/// Checks if user has provided their own header at `target/header.h`, otherwise
+/// we run cbindgen to generate one. Installs cffi if it's missing and we're inside a virtualenv
+///
+/// We're using the cffi recompiler, which reads the header, translates them into instructions
+/// how to load the shared library without the header and then writes those instructions to a
+/// file called `ffi.py`. This `ffi.py` will expose an object called `ffi`. This object is used
+/// in `__init__.py` to load the shared library into a module called `lib`.
+pub fn generate_cffi_declarations(crate_dir: &Path, python: &PathBuf) -> Result<String> {
+    let tempdir = tempdir()?;
+    let header = cffi_header(&crate_dir, &tempdir)?;
 
     let ffi_py = tempdir.as_ref().join("ffi.py");
 
     // Using raw strings is important because on windows there are path like
     // `C:\Users\JohnDoe\AppData\Local\TEmpl\pip-wheel-asdf1234` where the \U
-    // would otherwise be a broken unicode exscape sequence
+    // would otherwise be a broken unicode escape sequence
     let cffi_invocation = format!(
         r#"
 import cffi
@@ -403,9 +421,69 @@ recompiler.make_py_source(ffi, "ffi", r"{ffi_py}")
         header = header.display(),
     );
 
-    let output = Command::new(python)
-        .args(&["-c", &cffi_invocation])
-        .output()?;
+    let output = call_python(&python, &["-c", &cffi_invocation])?;
+    let install_cffi = if !output.status.success() {
+        // First, check whether the error was cffi not being installed
+        let last_line = str::from_utf8(&output.stderr)?.lines().last().unwrap_or("");
+        if last_line == "ModuleNotFoundError: No module named 'cffi'" {
+            // Then check whether we're running in a virtualenv.
+            // We don't want to modify any global environment
+            // https://stackoverflow.com/a/42580137/3549270
+            let output = call_python(
+                &python,
+                &["-c", "import sys\nprint(sys.base_prefix != sys.prefix)"],
+            )?;
+
+            match str::from_utf8(&output.stdout)?.trim() {
+                "True" => true,
+                "False" => false,
+                _ => {
+                    println!(
+                        "⚠  Failed to determine whether python at {:?} is running inside a virtualenv",
+                        &python
+                    );
+                    false
+                }
+            }
+        } else {
+            false
+        }
+    } else {
+        false
+    };
+
+    // If there was success or an error that was not missing cffi, return here
+    if !install_cffi {
+        return handle_cffi_call_result(&python, tempdir, &ffi_py, &output);
+    }
+
+    println!("⚠  cffi not found. Trying to install it");
+    // Call pip through python to don't do the wrong thing when python and pip
+    // are coming from different environments
+    let output = call_python(&python, &["-m", "pip", "install", "cffi"])?;
+    if !output.status.success() {
+        bail!(
+            "Installing cffi with `{:?} -m pip install cffi` failed: {}\n--- Stdout:\n{}\n--- Stderr:\n{}\n---\nPlease install cffi yourself.",
+            &python,
+            output.status,
+            str::from_utf8(&output.stdout)?,
+            str::from_utf8(&output.stderr)?
+        );
+    }
+    println!("🎁 Installed cffi");
+
+    // Try again
+    let output = call_python(&python, &["-c", &cffi_invocation])?;
+    handle_cffi_call_result(&python, tempdir, &ffi_py, &output)
+}
+
+/// Extracted into a function because this is needed twice
+fn handle_cffi_call_result(
+    python: &Path,
+    tempdir: TempDir,
+    ffi_py: &PathBuf,
+    output: &Output,
+) -> Result<String> {
     if !output.status.success() {
         bail!(
             "Failed to generate cffi declarations using {}: {}\n--- Stdout:\n{}\n--- Stderr:\n{}",
@@ -414,14 +492,14 @@ recompiler.make_py_source(ffi, "ffi", r"{ffi_py}")
             str::from_utf8(&output.stdout)?,
             str::from_utf8(&output.stderr)?,
         );
+    } else {
+        // Don't swallow warnings
+        std::io::stderr().write_all(&output.stderr)?;
+
+        let ffi_py_content = fs::read_to_string(&ffi_py)?;
+        tempdir.close()?;
+        Ok(ffi_py_content)
     }
-
-    // Don't swallow warnings
-    std::io::stderr().write_all(&output.stderr)?;
-
-    let ffi_py_content = fs::read_to_string(ffi_py)?;
-    tempdir.close()?;
-    Ok(ffi_py_content)
 }
 
 /// Copies the shared library into the module, which is the only extra file needed with bindings